% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_chromosome_blocks}
\alias{classify_chromosome_blocks}
\title{Delineate Syntenic Blocks via Sliding Window}
\usage{
classify_chromosome_blocks(chr_data, config)
}
\arguments{
\item{chr_data}{tibble. Spatial gene coordinates and initial \code{inferred_state_hmm1}.}

\item{config}{A \code{linguine_config} object.}
}
\value{
tibble. Consolidated coordinate blocks with finalized syntenic states.
}
\description{
Scans a chromosomal dataframe to partition the sequence into contiguous
syntenic blocks based on local density and purity thresholds.
}
