% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_linguine_config}
\alias{create_linguine_config}
\title{Create LINGUINE Configuration}
\usage{
create_linguine_config(
  dataset = "LINGUINE",
  min_chromosome_length_bp = 0,
  base_dir = getwd(),
  hmm_params = list(prob_self_syn_hmm1 = 0.995, prob_to_other_syn_hmm1 = 0.005,
    prob_emit_correct_syn_hmm1_value = 0.85, prob_emit_non_syn_obs_from_syn_hmm1_value =
    0.08, prob_emit_no_ortholog_from_syn_hmm1_value = 0.05,
    prob_emit_multiple_b_chrs_from_syn_hmm1_value = 0.01),
  thresholds = list(window_size = 10, purity_threshold = 0.8, min_final_block_size = 10,
    fisher_p = 0.01, fisher_odds = 2, min_ogs_by_block = 2, og_abundance = 90,
    fisher_p_lgs = 1e-20, fisher_odds_lgs = 4, paralogy_p = 1e-20, paralogy_odds = 10,
    outgroup_dominance = 0.05, parent_assignment = 0.6, ribbon_max_links = 5)
)
}
\arguments{
\item{dataset}{Character. Name of the dataset.}

\item{min_chromosome_length_bp}{Numeric. Minimum chromosome/scaffold length filter (in base pairs).}

\item{base_dir}{Character. Root directory for inputs/outputs. Defaults to current working directory.}

\item{hmm_params}{List. Transition and emission probabilities.}

\item{thresholds}{List. Thresholds for classification, graph clustering, and plotting.}
}
\value{
A structured list of class \code{linguine_config} containing all pipeline parameters and paths.
}
\description{
Generates a centralized configuration object for the Ancestral Genome Reconstruction Pipeline.
Instead of relying on global variables, this object is passed to downstream functions.
}
