% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_synteny.R
\name{extract_synteny_architecture}
\alias{extract_synteny_architecture}
\title{Extract Syntenic Architecture}
\usage{
extract_synteny_architecture(
  ref_species,
  comp_species,
  comparison_type,
  hmm_class_path,
  config
)
}
\arguments{
\item{ref_species}{Character. Reference lineage.}

\item{comp_species}{Character. Comparison lineage.}

\item{comparison_type}{Character. Architecture of the comparison.}

\item{hmm_class_path}{Character. Path to the HMM classification RDS.}

\item{config}{A \code{linguine_config} object.}
}
\value{
A list containing paths to the strict blocks and broad regions RDS files.
}
\description{
Consolidates Viterbi-inferred HMM classifications into contiguous
syntenic blocks and broad linkage groups.
}
