% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_architecture.R
\name{plot_architecture}
\alias{plot_architecture}
\title{Plot Ancestral Synteny Visualizations}
\usage{
plot_architecture(
  ref_species,
  comp_species,
  comparison_type,
  parent_node,
  config
)
}
\arguments{
\item{ref_species}{Character. Reference lineage.}

\item{comp_species}{Character. Comparison lineage.}

\item{comparison_type}{Character. Architecture of the comparison.}

\item{parent_node}{Character. Target internal node.}

\item{config}{A \code{linguine_config} object.}
}
\description{
Generates Oxford Grids and high-fidelity linear/ribbon plots mapping
algorithmic ancestral states to extant sequences.
}
