% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_synteny.R
\name{plot_synteny_summary}
\alias{plot_synteny_summary}
\title{Plot Syntenic Summary Statistics}
\usage{
plot_synteny_summary(ref_species, comp_species, comparison_type, config)
}
\arguments{
\item{ref_species}{Character. Reference lineage.}

\item{comp_species}{Character. Comparison lineage.}

\item{comparison_type}{Character. Architecture of the comparison.}

\item{config}{A \code{linguine_config} object.}
}
\value{
Invisible NULL. Saves plots to the configured plots directory.
}
\description{
Generates visual distributions and karyoplots to map algorithmic
blocks back to physical scaffolds.
}
