% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_synteny_data.R
\name{prepare_synteny_data}
\alias{prepare_synteny_data}
\title{Prepare Genomic and Orthology Data for HMM}
\usage{
prepare_synteny_data(
  ref_species,
  comp_species,
  comparison_type,
  parent_node,
  comp_node_daughters,
  config
)
}
\description{
Prepare Genomic and Orthology Data for HMM
}
