% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_node.R
\name{reconstruct_node}
\alias{reconstruct_node}
\title{Reconstruct Ancestral Node Topology}
\usage{
reconstruct_node(
  ref_species,
  comp_species,
  comparison_type,
  parent_node,
  species_tree,
  config
)
}
\arguments{
\item{ref_species}{Character. Reference lineage.}

\item{comp_species}{Character. Comparison lineage.}

\item{comparison_type}{Character. Architecture of the comparison.}

\item{parent_node}{Character. Target internal node being reconstructed.}

\item{species_tree}{phylo object. The full phylogenetic tree for outgroup polarization.}

\item{config}{A \code{linguine_config} object.}
}
\value{
Path to the saved ancestral genome RDS file.
}
\description{
Executes topological graph clustering to merge, resolve, and polarize
ancestral linkage groups using bipartite matching and an outgroup.
}
