% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_synteny_hmm.R
\name{run_synteny_hmm}
\alias{run_synteny_hmm}
\title{Execute Synteny HMM Inference}
\usage{
run_synteny_hmm(
  ref_species,
  comp_species,
  comparison_type,
  hmm_obs_path,
  config
)
}
\arguments{
\item{ref_species}{Character. Reference lineage.}

\item{comp_species}{Character. Comparison lineage.}

\item{comparison_type}{Character. Architecture of the comparison.}

\item{hmm_obs_path}{Character. Path to the RDS file generated by prepare_synteny_data.}

\item{config}{A \code{linguine_config} object.}
}
\value{
Character path to the saved HMM classification RDS file.
}
\description{
Configures transition/emission matrices dynamically based on input
dimensions and applies the Viterbi algorithm to decode hidden ancestral states.
}
